Added "raw" group to reduce size of symbol table, further GetSymbolGroup optimizations, changed block size to 32-bit, other bug fixes

+++ Precision.h
@@ -6,5 +6,5 @@ typedef uint16_t prob_t;
 typedef uint16_t symidx_t;
 typedef uint16_t group_t;
 typedef uint32_t count_t;
-typedef uint16_t block_t;
+typedef uint32_t block_t;
 typedef uint64_t state_t;
